Create ImportingDataset
There are a total of Thirty-two attributes in the dataset, of which thirty will be input features, one will be for classification and one attribute denotes the ID number. Inputs are features and measurements of the tumor. Some of these parameters include measurements of the nuclei of cells in tumor-like the radius, texture, perimeter, and area. The output parameter will be the classification, “m” denotes malignant, and “b” is for benign. In this report, “b” will be represented as 0 and “m” will be represented as 1. The dataset is taken from the UCI Machine Learning Repository.