Create BuildingModel.py
Training and testing data is split using the train_test_split. 80% of data is used for training whereas 20% of the data is used for testing. Convolutional neural networks only accept three-dimensional data and we currently have x and y in two-dimensional so data is reshaped into three-dimensional by converting it to a NumPy array and then applying standardization.
For building the model a number of epochs have to be defined so 50 epochs are defined and a sequential model is stated. Further convolution 1D with filter size 32, kernel size 2, activation function set to “relu”. For the first layer of the convolutional neural network, an input shape is provided of 30 columns and 1 row. Batch normalization is added to reduce variance in data when it is shifting from one layer to another. Dropout is also added and 20 percent of neurons are randomly dropped.
 
For the second layer, the input shape function is removed and the filter size is set to 64. Neuron drop using dropout is done for 50 percent of the neurons.
 So a two-layer convolutional network is formed and this two-dimensional data is flattened into a vector. Further, a dense layer is provided with 64 units, activation function called is “relu”, dropout is done for 50 percent.
The final output is called which is the dense layer and the activation function it uses is sigmoid.