Create CompileModel.py
For compilation, an optimizer adam is passed and the learning rate is set to 0.00005 which is a very slow learning rate. The loss function will be binary cross-entropy and the ‘accuracy’ matrix is called. The history of model training is recorded and the model is fitted with both training and testing dataset along with epochs is passed, verbose is set to 1. When the code is run the training of the model begins. The accuracy starts from 67 percent and goes till 96 percent which is considered really good. Whereas the validation accuracy starts at 84 percent and goes to 95 percent.
The plot of a learning curve and validation accuracy values are printed for 50 epochs. Values of accuracy, validation accuracy, loss, and validation loss are plotted. Validation accuracy is always greater than the training accuracy therefore the model is not overfitting and validation loss is lower than training loss so the model can be kept on training. Accuracy is the inverse of the cost therefore a good model should have accuracy as large as possible so that the cost is kept minimum.